fix overflowing link text
