+++ packages/components/src/components/Dropdown/Dropdown.stories.tsx
@@ -63,6 +63,37 @@ export const Default: Story = {
     ),
 };
 
+export const LinkItems: Story = {
+    render: ({ ...args }) => (
+        <Dropdown.Root {...args}>
+            <Dropdown.Trigger>
+                <Button>Trigger</Button>
+            </Dropdown.Trigger>
+            <Dropdown.Content>
+                <Dropdown.Item onSelect={() => {}}>
+                    <a href="https://frontify.com">
+                        Link 1
+                        {/* <Dropdown.Slot name="right">
+                            <IconIcon size={16} />
+                        </Dropdown.Slot> */}
+                    </a>
+                </Dropdown.Item>
+                <Dropdown.Item onSelect={() => {}}>
+                    <a href="https://frontify.com">
+                        Link 2
+                        <Dropdown.Slot name="left">
+                            <IconIcon size={16} />
+                        </Dropdown.Slot>
+                    </a>
+                </Dropdown.Item>
+                <Dropdown.Item onSelect={() => {}}>
+                    <span>Item 3</span>
+                </Dropdown.Item>
+            </Dropdown.Content>
+        </Dropdown.Root>
+    ),
+};
+
 export const DisabledItems: Story = {
     render: ({ ...args }) => (
         <Dropdown.Root {...args}>
@@ -162,7 +193,12 @@ export const OverflowingText: Story = {
             <Dropdown.Content>
                 <Dropdown.Item onSelect={() => {}}>Item 1</Dropdown.Item>
                 <Dropdown.Item onSelect={() => {}}>
-                    Item 2 - This is a very long text that overflows the element it is contained in
+                    <a href="https://frontify.com">
+                        Link 2 - This is a very long text that overflows the element it is contained in
+                        <Dropdown.Slot name="right">
+                            <IconIcon size={16} />
+                        </Dropdown.Slot>
+                    </a>
                 </Dropdown.Item>
                 <Dropdown.Item onSelect={() => {}}>Item 3</Dropdown.Item>
                 <Dropdown.SubMenu>
@@ -370,33 +406,6 @@ export const WithTooltip: Story = {
                 <Dropdown.Item onSelect={() => {}}>Item 1</Dropdown.Item>
                 <Dropdown.Item onSelect={() => {}}>Item 2</Dropdown.Item>
                 <Dropdown.Item onSelect={() => {}}>Item 3</Dropdown.Item>
-            </Dropdown.Content>
-        </Dropdown.Root>
-    ),
-};
-
-export const LinkItems: Story = {
-    render: ({ ...args }) => (
-        <Dropdown.Root {...args}>
-            <Dropdown.Trigger>
-                <Button>Trigger</Button>
-            </Dropdown.Trigger>
-            <Dropdown.Content>
-                <Dropdown.Item onSelect={() => {}}>
-                    <a href="https://frontify.com">Link 1</a>
-                    <Dropdown.Slot name="right">
-                        <IconIcon size={16} />
-                    </Dropdown.Slot>
-                </Dropdown.Item>
-                <Dropdown.Item onSelect={() => {}}>
-                    <a href="https://frontify.com">Link 2</a>
-                    <Dropdown.Slot name="left">
-                        <IconIcon size={16} />
-                    </Dropdown.Slot>
-                </Dropdown.Item>
-                <Dropdown.Item onSelect={() => {}}>
-                    <span>Item 3</span>
-                </Dropdown.Item>
             </Dropdown.Content>
         </Dropdown.Root>
     ),
